# api/openapi.yaml
openapi: 3.1.0
info:
  title: 127ohoh1 control-plane API (development reference)
  version: 0.1.0
  description: >
    Public reference API. Paths are added as phases land. Payment endpoints are
    simulated, development-only, and fail closed unless explicitly enabled.
    All timestamps are UTC RFC 3339. Errors use the envelope in `Error`.
servers:
  - url: http://localhost:8080
security:
  - session: []
paths:
...
  /v1/auth/challenges:
    post:
      security: []
      summary: Request a single-use, short-lived challenge for a public key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [public_key]
              additionalProperties: false
              properties:
                public_key: { type: string, description: "Unpadded base64url Ed25519 public key (32 bytes)" }
      responses:
        "201":
          description: Challenge
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Challenge' }
        "400": { $ref: '#/components/responses/Error' }
        "429": { $ref: '#/components/responses/Error' }
  /v1/auth/sessions:
    post:
      security: []
      summary: Exchange a signed challenge for a short-lived session token
      description: >
        The signature covers the canonical bytes defined in docs/tunnel-protocol.md
        (domain-separated, length-prefixed). The challenge is consumed even if
        verification fails.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge_id, nonce, signature]
              additionalProperties: false
              properties:
                challenge_id: { type: string }
                nonce: { type: string }
                signature: { type: string }
      responses:
        "201":
          description: Session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        "401": { $ref: '#/components/responses/Error' }
        "403": { $ref: '#/components/responses/Error' }
  /v1/endpoints:
    get:
      summary: List the caller's live endpoints
      responses:
        "200":
          description: Endpoints
          content:
            application/json:
              schema:
                type: object
                properties:
                  endpoints: { type: array, items: { $ref: '#/components/schemas/Endpoint' } }
    post:
      summary: Create a random free endpoint
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                kind: { type: string, enum: [random] }
      responses:
        "201":
          description: Created (or replayed via Idempotent-Replayed header)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Endpoint' }
        "409": { $ref: '#/components/responses/Error' }
        "429": { $ref: '#/components/responses/Error' }
  /v1/endpoints/{endpoint_id}:
    parameters:
      - { name: endpoint_id, in: path, required: true, schema: { type: string } }
    get:
      summary: Get an owned endpoint
      responses:
        "200":
          description: Endpoint
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Endpoint' }
        "404": { $ref: '#/components/responses/Error' }
    delete:
      summary: Release an owned endpoint (idempotent)
      responses:
        "200":
          description: Released endpoint
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Endpoint' }
        "404": { $ref: '#/components/responses/Error' }
  /v1/endpoints/{endpoint_id}/tunnel-credentials:
    parameters:
      - { name: endpoint_id, in: path, required: true, schema: { type: string } }
    post:
      summary: Issue a short-lived tunnel credential scoped to one endpoint
      responses:
        "201":
          description: Credential
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TunnelCredential' }
        "409": { $ref: '#/components/responses/Error' }
        "410": { $ref: '#/components/responses/Error' }
...
  /v1/usage/batches:
    post:
      summary: "Edge service: report usage (idempotent per edge_id + batch_id)"
      security: [ { edge: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [edge_id, batch_id, items]
              properties:
                edge_id: { type: string }
                batch_id: { type: string }
                items:
                  type: array
                  maxItems: 1000
                  items:
                    type: object
                    properties:
                      endpoint_id: { type: string }
                      period_start: { type: string, format: date-time, description: "first instant of a UTC month" }
                      bytes: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Applied (Idempotent-Replayed header on duplicates)
        "400": { $ref: '#/components/responses/Error' }
        "401": { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    session:
      type: http
      scheme: bearer
    edge:
      type: http
      scheme: bearer
      description: Edge service credential (constant-time compared)
    admin:
      type: http
      scheme: bearer
      description: Administrative token; the admin API is absent (404) when none is configured
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 8-128 printable ASCII characters. Reuse with a different request returns 409 idempotency_key_reused.
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            details: { type: object }
...
    Challenge:
      type: object
      properties:
        challenge_id: { type: string }
        nonce: { type: string, description: "base64url, at least 128 bits" }
        audience: { type: string }
        protocol_version: { type: integer }
        issued_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
    Session:
      type: object
      properties:
        session_token: { type: string }
        expires_at: { type: string, format: date-time }
        principal_id: { type: string }
        fingerprint: { type: string }
    Endpoint:
      type: object
      properties:
        id: { type: string }
        hostname: { type: string }
        url: { type: string }
        kind: { type: string, enum: [random, reserved] }
        state: { type: string, enum: [PENDING, ACTIVE, SUSPENDED, EXPIRED, RELEASED] }
        plan: { type: string }
        created_at: { type: string, format: date-time }
        last_active_at: { type: string, format: date-time }
    TunnelCredential:
      type: object
      properties:
        token: { type: string }
        expires_at: { type: string, format: date-time }
        endpoint_id: { type: string }
        hostname: { type: string }
        url: { type: string }
        edge:
          type: object
          properties:
            addr: { type: string }
            audience: { type: string }
...